src/test/java/com/skplanet/plalab/ez2sugul/JadenCaseTest.java
package com.skplanet.plalab.ez2sugul;

import static org.junit.Assert.*;

import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;

public class JadenCaseTest {
	
	JadenCase jadenCase = new JadenCase();

	@BeforeClass
	public static void setUpBeforeClass() throws Exception {
	}

	@AfterClass
	public static void tearDownAfterClass() throws Exception {
	}

	@Before
	public void setUp() throws Exception {
	}

	@After
	public void tearDown() throws Exception {
	}

	@Test
	public void test() {
		assertEquals("Most Trees Are Blue", jadenCase.toJadenCase("most trees are blue"));
	}
	
	@Test
	public void nullTest() {
		assertEquals(null, jadenCase.toJadenCase(""));
	}

}
